Make use of plugin (intercepter) concept in Magento
