+++ src/etc/di.xml
@@ -12,28 +12,49 @@
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
         xsi:noNamespaceSchemaLocation="urn:magento:framework:ObjectManager/etc/config.xsd">
     <!-- replacing core preferences -->
+
+    <type name="\Magento\Sales\Model\Order\Email\Sender\OrderSender">
+        <plugin name="magento_sales_model_order_email_sender_ordersender_plugin"
+                type="Fooman\EmailAttachments\Model\Email\Sender\Plugin"
+                sortOrder="10"
+                disabled="false"/>
+    </type>
+
+    <type name="\Magento\Sales\Model\Order\Email\Sender\InvoiceSender">
+        <plugin name="magento_sales_model_order_email_sender_invoicesender_plugin"
+                type="Fooman\EmailAttachments\Model\Email\Sender\Plugin"
+                sortOrder="10"
+                disabled="false"/>
+    </type>
+
+    <type name="\Magento\Sales\Model\Order\Email\Sender\ShipmentSender">
+        <plugin name="magento_sales_model_order_email_sender_shipmentsender_plugin"
+                type="Fooman\EmailAttachments\Model\Email\Sender\Plugin"
+                sortOrder="10"
+                disabled="false"/>
+    </type>
+
+    <type name="\Magento\Sales\Model\Order\Email\Sender\CreditmemoSender">
+        <plugin name="magento_sales_model_order_email_sender_creditmemosender_plugin"
+                type="Fooman\EmailAttachments\Model\Email\Sender\Plugin"
+                sortOrder="10"
+                disabled="false"/>
+    </type>
+
     <preference for="\Magento\Framework\Mail\Template\TransportBuilder"
                 type="Fooman\EmailAttachments\Model\MailTransportBuilder"/>
     <preference for="\Magento\Sales\Model\Order\Email\SenderBuilder"
                 type="Fooman\EmailAttachments\Model\SenderBuilder"/>
 
-    <preference for="\Magento\Sales\Model\Order\Email\Sender\OrderSender"
-                type="Fooman\EmailAttachments\Model\Email\Sender\OrderSender"/>
     <preference for="\Magento\Sales\Model\Order\Email\Sender\OrderCommentSender"
                 type="Fooman\EmailAttachments\Model\Email\Sender\OrderCommentSender"/>
 
-    <preference for="\Magento\Sales\Model\Order\Email\Sender\InvoiceSender"
-                type="Fooman\EmailAttachments\Model\Email\Sender\InvoiceSender"/>
     <preference for="\Magento\Sales\Model\Order\Email\Sender\InvoiceCommentSender"
                 type="Fooman\EmailAttachments\Model\Email\Sender\InvoiceCommentSender"/>
 
-    <preference for="\Magento\Sales\Model\Order\Email\Sender\ShipmentSender"
-                type="Fooman\EmailAttachments\Model\Email\Sender\ShipmentSender"/>
     <preference for="\Magento\Sales\Model\Order\Email\Sender\ShipmentCommentSender"
                 type="Fooman\EmailAttachments\Model\Email\Sender\ShipmentCommentSender"/>
 
-    <preference for="\Magento\Sales\Model\Order\Email\Sender\CreditmemoSender"
-                type="Fooman\EmailAttachments\Model\Email\Sender\CreditmemoSender"/>
     <preference for="\Magento\Sales\Model\Order\Email\Sender\CreditmemoCommentSender"
                 type="Fooman\EmailAttachments\Model\Email\Sender\CreditmemoCommentSender"/>
